feat!: Add common library for WF and non WF code
