+++ packages/common/README.md
@@ -1,0 +1,7 @@
+# `@temporalio/common`
+
+[![NPM](https://img.shields.io/npm/v/@temporalio/common?style=for-the-badge)](https://www.npmjs.com/package/@temporalio/common)
+
+Part of the [Temporal](https://temporal.io) [Node.js SDK](https://www.npmjs.com/package/temporalio).
+
+See documentation on the [documentation site](https://docs.temporal.io/docs/node/introduction).
