+++ packages/common/src/encoding.ts
