+++ packages/proto/package.json
@@ -17,9 +17,11 @@
   "dependencies": {
     "@temporalio/core-bridge": "file:../core-bridge",
     "@types/long": "^4.0.1",
-    "glob": "^7.1.6",
     "long": "^4.0.0",
     "protobufjs": "^6.10.2"
+  },
+  "devDependencies": {
+    "glob": "^7.1.6"
   },
   "bugs": {
     "url": "https://github.com/temporalio/sdk-node/issues"
