+++ packages/test/src/perf/setup.ts
@@ -1,6 +1,6 @@
 import arg from 'arg';
 import { Connection } from '@temporalio/client';
-import { msToTs } from '@temporalio/workflow/lib/time';
+import { msToTs } from '@temporalio/common';
 import { SetupArgSpec, setupArgSpec, getRequired } from './args';
 
 async function createNamespace(connection: Connection, namespace: string, maxAttempts = 100, retryIntervalSecs = 1) {
