+++ packages/test/src/test-time.ts
@@ -1,6 +1,6 @@
 import test from 'ava';
 import Long from 'long';
-import { msToTs } from '@temporalio/workflow/lib/time';
+import { msToTs } from '@temporalio/common';
 
 test('msToTs converts to Timestamp', (t) => {
   t.deepEqual({ seconds: Long.fromInt(600), nanos: 0 }, msToTs('10 minutes'));
