+++ packages/test/src/test-workflows.ts
@@ -5,8 +5,7 @@ import Long from 'long';
 import dedent from 'dedent';
 import { coresdk } from '@temporalio/proto';
 import { ApplyMode } from '@temporalio/workflow';
-import { defaultDataConverter } from '@temporalio/workflow/lib/converter/data-converter';
-import { msToTs } from '@temporalio/workflow/lib/time';
+import { defaultDataConverter, msToTs } from '@temporalio/common';
 import { Workflow } from '@temporalio/worker/lib/workflow';
 import { WorkflowIsolateBuilder } from '@temporalio/worker/lib/isolate-builder';
 import { DefaultLogger } from '@temporalio/worker/lib/logger';
