+++ packages/worker/tsconfig.json
@@ -5,7 +5,11 @@
     "outDir": "./lib",
     "rootDir": "./src"
   },
-  "references": [{ "path": "../workflow/tsconfig.json" }, { "path": "../activity/tsconfig.json" }],
+  "references": [
+    { "path": "../client/tsconfig.json" },
+    { "path": "../workflow/tsconfig.json" },
+    { "path": "../activity/tsconfig.json" }
+  ],
   "include": ["./src/**/*.ts"],
   "exclude": ["node_modules"]
 }
